dna2vec/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dna2vec"
version = "0.0.4"
description = "A package for sequence alignment of single reads"
authors = []

classifiers = [
    "Development Status :: 1 - Pre-alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]

requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "umap-learn>=0.1.1",
    "tqdm>=4.62.0",
    "pinecone-client[grpc]>2.2.0",
    "pydantic>=1.8.2,<2.0.0",
    "typer>=0.3.2",
    "tokenizers>=0.13.3",
    "wandb>=0.12.1",
    "biopython>=1.81",
    "pysam>=0.21.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.0.261", 
  "black[jupyter]>=23.3.0",
]
tests = [
  "pytest>=7.1.2",
  "pytest-cov>=3.0.0",
]


[project.readme]
file = "readme.md"
content-type = "text/markdown"

[tool.setuptools.package-data]
"*" = ["*.csv"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.coverage.run]
omit = [
    "**/tests/*",
    "**/about.py",
    "**/dev/*",
]

[tool.mypy]
ignore_missing_imports = true
no_implicit_optional = true
warn_unreachable = true

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
  "A",
  "ANN",
  "ARG",
  "B",
  "C4",
  "COM",
  "D417",
  "E",
  "ERA",
  "F",
  "I",
  "ICN",
  "NPY001",
  "PD002",
  "PIE",
  "PLE",
  "PLW",
  "PT",
  "UP",
  "Q",
  "PTH",
  "RSE",
  "RET",
  "RUF",
  "SIM",
  "W"
]
ignore = [
  "ANN101",
  "ANN401",
  "E402",
  "E501",
  "F401",
  "F841",
  "RET504",
  "ANN202",
]
ignore-init-module-imports = true
# Allow autofix for all enabled rules (when `--fix`) is provided.
unfixable = ["ERA"]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "__init__.py",
  ".venv",
  ".env",
  ".git",
  "__pycache__",
  "dev/**",
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py38"

[tool.ruff.flake8-annotations]
mypy-init-return = true
suppress-none-returning = true

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.setuptools]
include-package-data = true
